# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dispatch API
  version: 3.0.0
  description: |
    Open-source bulk email campaign platform with tracking, multi-provider support, and batch processing.

    ## Authentication
    Most endpoints require authentication via a `session_token` cookie obtained from the login endpoint.
    Alternatively, an API key can be passed via the `X-API-Key` header.

    ## CSRF Protection
    All mutation endpoints (POST, PUT, DELETE) require a valid `X-CSRF-Token` header.
    The CSRF token is issued as a cookie on every response.

    ## Common Response Format
    All endpoints return a standard envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "message": "Human-readable message",
      "meta": { "pagination": { ... } }
    }
    ```
  contact:
    name: Dispatch
    url: https://github.com/iamjpsingh/bulk-email-sender
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session token cookie set on login/register.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token required for POST/PUT/DELETE requests. Value is read from the `csrf_token` cookie.

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (shape varies per endpoint).
        message:
          type: string
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Something went wrong
        error:
          type: string
      required: [success, message]
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true

    # -- Auth --
...
    ContactList:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        contact_count:
          type: integer
        created_at:
          type: string
          format: date-time
    Contact:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        company:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, unsubscribed, bounced, complained]
        tags:
          type: string
        custom_fields:
          type: object
        created_at:
          type: string
          format: date-time
    ContactCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        company:
          type: string
        phone:
          type: string
        tags:
          type: string
        custom_fields:
          type: object
      required: [email]

    # -- Templates --
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
...
    SearchParam:
      name: search
      in: query
      schema:
        type: string

# ---------------------------------------------------------------------------
# Default Security
# ---------------------------------------------------------------------------
security:
  - cookieAuth: []
  - apiKeyAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /contacts/lists:
    get:
      tags: [Contacts]
      summary: List contact lists
      operationId: listContactLists
      responses:
        '200':
          description: Array of contact lists.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          lists:
                            type: array
                            items:
                              $ref: '#/components/schemas/ContactList'

    post:
      tags: [Contacts]
      summary: Create a contact list
      operationId: createContactList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required: [name]
      responses:
        '200':
          description: Contact list created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ContactList'
        '400':
          description: Name is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contacts/lists/{id}:
    put:
      tags: [Contacts]
      summary: Update a contact list
      operationId: updateContactList
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: List updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: List not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Contacts]
      summary: Delete a contact list
      operationId: deleteContactList
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: List not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contacts/search:
    get:
      tags: [Contacts]
      summary: Search contacts across all lists
      operationId: searchContacts
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        '200':
          description: Matching contacts.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          contacts:
                            type: array
                            items:
                              $ref: '#/components/schemas/Contact'
  /contacts/{listId}:
    get:
      tags: [Contacts]
      summary: List contacts in a list
      operationId: listContacts
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/SearchParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, unsubscribed, bounced, complained]
        - name: tags
          in: query
          schema:
            type: string
            description: Comma-separated tag names.
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated contacts.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Contact'
                      meta:
                        type: object
                        properties:
                          pagination:
                            $ref: '#/components/schemas/Pagination'

    post:
      tags: [Contacts]
      summary: Add a contact to a list
      operationId: addContact
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '200':
          description: Contact added.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Contact'
        '400':
          description: Email is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email in this list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contacts/item/{id}:
    put:
      tags: [Contacts]
      summary: Update a contact
      operationId: updateContact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '200':
          description: Contact updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Contact not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contacts/bulk/delete:
    post:
      tags: [Contacts]
      summary: Bulk delete contacts
      operationId: bulkDeleteContacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
              required: [ids]
      responses:
        '200':
          description: Contacts deleted.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: integer
  /contacts/bulk/tag:
    post:
      tags: [Contacts]
      summary: Bulk tag contacts
      operationId: bulkTagContacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
              required: [ids, tags]
      responses:
        '200':
          description: Contacts tagged.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          updated:
                            type: integer
  /contacts/bulk/move:
    post:
      tags: [Contacts]
      summary: Bulk move contacts to another list
      operationId: bulkMoveContacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
                target_list_id:
                  type: string
              required: [ids, target_list_id]
      responses:
        '200':
          description: Contacts moved.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          moved:
                            type: integer
  /contacts/{listId}/import:
    post:
      tags: [Contacts]
      summary: Import contacts from a file
      operationId: importContacts
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Excel or CSV file with contacts.
                fieldMapping:
                  type: string
                  description: JSON object mapping file columns to contact fields.
                skipDuplicates:
                  type: string
                  enum: ['true', 'false']
                  default: 'true'
              required: [file]
      responses:
        '200':
          description: Import result with counts.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          imported:
                            type: integer
                          duplicates:
                            type: integer
                          invalid:
                            type: integer
                          errors:
                            type: array
                            items:
                              type: string
...
  /contacts/validate:
    post:
      tags: [Contacts]
      summary: Validate a batch of emails
      operationId: validateEmails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emails:
                  type: array
                  items:
                    type: string
                    format: email
                  maxItems: 100
              required: [emails]
      responses:
        '200':
          description: Validation results per email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /contacts/validate-single:
    post:
      tags: [Contacts]
      summary: Validate a single email
      operationId: validateSingleEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '200':
          description: Validation result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # =========================================================================
  # Templates
  # =========================================================================
...